# docs/web/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Admin REST API шины данных (§7 ТЗ). Сессии в Redis +\nAPI-токены (Bearer db_*). SPA по embed.FS отдаётся\nфолбэком на index.html для всех путей не из /api/.",
        "title": "Nexus Web API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/allowed-hosts": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Каталог разрешённых хостов: поиск/листинг (§23).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "поиск по паттерну и описанию",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "фильтр по типу: exact|wildcard|regex",
                        "name": "kind",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "лимит (по умолчанию 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ListHostsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Идемпотентно по case-insensitive паттерну: повтор вернёт существующую запись.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Создать разрешённый хост (admin, §23).",
                "parameters": [
                    {
                        "description": "host",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.hostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.hostResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/allowed-hosts/preview": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Превью паттерна: какие URL разрешит/заблокирует (§23).",
                "parameters": [
                    {
                        "description": "preview",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.hostPreviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.HostPreviewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/allowed-hosts/{id}": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Нельзя удалить хост, используемый узлами (409).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Удалить разрешённый хост (admin, §23).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "host id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Описание меняется всегда; паттерн/тип — только если хост не используется (usage_count=0).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Изменить разрешённый хост (admin, §23).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "host id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "host",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.hostRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/nodes/{id}/allowed-hosts": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Хосты, привязанные к узлу (§23).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ListHostsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Привязать хост к узлу (admin, §23).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "host_id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.attachHostRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/nodes/{id}/allowed-hosts/{host_id}": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "allowed-hosts"
                ],
                "summary": "Отвязать хост от узла (admin, §23).",
                "parameters": [
                    {
                        "type": "string",
                        "description": "node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "host id",
                        "name": "host_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_web_adapter_in_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_web_adapter_in_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "password_change_required"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
        "internal_web_adapter_in_http.HostPreviewResponse": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "blocked": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_web_adapter_in_http.ListHostsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_web_adapter_in_http.hostResponse"
                    }
                }
            }
        },
...
        "internal_web_adapter_in_http.attachHostRequest": {
            "type": "object",
            "properties": {
                "host_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_web_adapter_in_http.hostPreviewRequest": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "pattern": {
                    "type": "string"
                },
                "test_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_web_adapter_in_http.hostRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "pattern": {
                    "type": "string"
                }
            }
        },
        "internal_web_adapter_in_http.hostResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "pattern": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiTokenAuth": {
            "description": "\"Bearer db_\u003c32-байт-base64\u003e\". См. §7.14.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "nexus_session",
            "in": "cookie"
        }
    }
}